Fancier boss info bars for Naga and Lich!
Naga has 12 segments for both body and boss bar.

Lich has 6 segments and is yellow when it has shields. Bar loses a segment when it loses a shield. Bar refills and is a regular progression bar in color purple when it loses all shields. Bar turns red when it enters phase 3 when 6 of its loyal zombies are slain. Also emits   A N G E R Y   particles!