Fixed up code for the Uberous Soil, fixed a crash
The game would freeze if you tried to plant a mushroom onto the soil, but it failed to grow.
The soil now turns into a grass block if you grow a grass on it,
turns into mycelium if you grow a mushroom on it.
Turning into mycelium helps the mushroom actually grow in any light condition, making the process a lot easier on the player.