Trollsteinn fixed, now has comparator outputs
I realize random ticks aren't nearly as smooth as the instant updates Trollsteinn used to have, but this is the least laggy solution I could think of. If you have a better one, please feel free to implement it.

The block also now has a comparator output based on the highest light level in it's vicinity.

Also fixed a typo