Fixed ChainBlock sometimes failing to break blocks
Made the AABB center around the impact point instead of the center of the Chain Block itself. Due to how projectiles work, the chain block would sometimes be half a block away from the impact point, making it not break any blocks, due to the logic checking for it's position, instead of the impact point. This would mostly happen when hitting the ground